fix(key-card): fix types
fixed types

WP-722

TICKET: WP-722
